entityパッケージ名の変更
