+++ itemsystem/src/main/java/com/example/itemsystem/detailService/DetailService.java
@@ -8,7 +8,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.example.itemsystem.entiry.AdminEntity;
+import com.example.itemsystem.entity.AdminEntity;
 import com.example.itemsystem.repository.AdminRepository;
 
 @Service
