+++ itemsystem/src/main/java/com/example/itemsystem/entity/AdminEntity.java
@@ -1,4 +1,4 @@
-package com.example.itemsystem.entiry;
+package com.example.itemsystem.entity;
 
 import java.time.LocalDateTime;
