+++ itemsystem/src/main/java/com/example/itemsystem/entity/ItemUnderCategoryEntity.java
@@ -1,4 +1,4 @@
-package com.example.itemsystem.entiry;
+package com.example.itemsystem.entity;
 
 import java.time.LocalDateTime;
 
@@ -7,6 +7,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
@@ -14,13 +16,16 @@ import lombok.Data;
 
 @Entity
 @Data
-@Table(name = "item_large_category")
-public class ItemLargeCategoryEntity {
-	
+@Table(name = "item_under_category")
+public class ItemUnderCategoryEntity {
 	@Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
+	
+	@ManyToOne
+	@JoinColumn(name = "item_large_category_id", referencedColumnName = "id")
+    private ItemLargeCategoryEntity itemlargeCategory;
 	
 	@Column(name = "name")
 	private String name;
