+++ itemsystem/src/main/java/com/example/itemsystem/entity/ManufacturesEntity.java
@@ -1,4 +1,4 @@
-package com.example.itemsystem.entiry;
+package com.example.itemsystem.entity;
 
 import java.time.LocalDateTime;
 
@@ -7,25 +7,21 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import lombok.Data;
 
+
 @Entity
 @Data
-@Table(name = "item_small_category")
-public class ItemSmallCategoryEntity {
+@Table(name = "manufactures")
+public class ManufacturesEntity {
+	
 	@Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
-	
-	@ManyToOne
-	@JoinColumn(name = "item_under_category_id", referencedColumnName = "id")
-    private ItemUnderCategoryEntity itemUnderCategory;
 	
 	@Column(name = "name")
 	private String name;
