+++ itemsystem/src/main/java/com/example/itemsystem/repository/AdminRepository.java
@@ -2,7 +2,7 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.AdminEntity;
+import com.example.itemsystem.entity.AdminEntity;
 
 public interface AdminRepository extends JpaRepository<AdminEntity, Long>{
 	AdminEntity findByEmail(String email);
