+++ itemsystem/src/main/java/com/example/itemsystem/repository/ItemLargeCategoryRepository.java
@@ -2,7 +2,7 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.ItemLargeCategoryEntity;
+import com.example.itemsystem.entity.ItemLargeCategoryEntity;
 
 public interface ItemLargeCategoryRepository extends JpaRepository<ItemLargeCategoryEntity, Long>{
