+++ itemsystem/src/main/java/com/example/itemsystem/repository/ItemSmallCategoryRepository.java
@@ -2,8 +2,10 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.ItemSmallCategoryEntity;
+import com.example.itemsystem.entity.ItemSmallCategoryEntity;
 
 public interface ItemSmallCategoryRepository extends JpaRepository<ItemSmallCategoryEntity, Long>{
+
+
 
 }
