+++ itemsystem/src/main/java/com/example/itemsystem/repository/ItemUnderCategoryRepository.java
@@ -2,8 +2,10 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.ItemUnderCategoryEntity;
+import com.example.itemsystem.entity.ItemUnderCategoryEntity;
 
 public interface ItemUnderCategoryRepository extends JpaRepository<ItemUnderCategoryEntity, Long>{
+
+
 
 }
