+++ itemsystem/src/main/java/com/example/itemsystem/repository/ManufactresRepository.java
@@ -2,7 +2,7 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.ManufacturesEntity;
+import com.example.itemsystem.entity.ManufacturesEntity;
 
 public interface ManufactresRepository extends JpaRepository<ManufacturesEntity, Long>{
