+++ itemsystem/src/main/java/com/example/itemsystem/repository/PermissionRepository.java
@@ -2,7 +2,7 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.PermissionEntity;
+import com.example.itemsystem.entity.PermissionEntity;
 
 
 public interface PermissionRepository extends JpaRepository<PermissionEntity, Long> {
