+++ itemsystem/src/main/java/com/example/itemsystem/repository/RoleRepository.java
@@ -2,7 +2,7 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.RoleEntity;
+import com.example.itemsystem.entity.RoleEntity;
 
 public interface RoleRepository  extends JpaRepository<RoleEntity, Long> {
