+++ itemsystem/src/main/java/com/example/itemsystem/repository/ShopRepository.java
@@ -2,7 +2,7 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entiry.ShopEntity;
+import com.example.itemsystem.entity.ShopEntity;
 
 public interface ShopRepository extends JpaRepository<ShopEntity, Long> {
