+++ itemsystem/src/main/resources/templates/top.html
@@ -10,7 +10,9 @@
     <div class="container">
         <h1>TOPメニュー</h1>
         <div class="menu">
-            <button class="menu-button">商品一覧</button>
+			<form action="/item/list" method="post"th:action="@{/item/list}">
+				<button class="menu-button">商品一覧</button>
+			</form>
             <button class="menu-button">店舗一覧</button>
 			<form action="/admin/list" method="post"th:action="@{/admin/list}">
 				<button class="menu-button">管理者一覧</button>
